% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetero_check.R
\name{hetero_check}
\alias{hetero_check}
\title{Breusch-Pagan Test for Heteroscedasticity}
\usage{
hetero_check(model)
}
\arguments{
\item{model}{A fitted linear model object.}
}
\value{
A data frame containing the BP statistic,
degrees of freedom, p-value, and conclusion.
}
\description{
Performs the Breusch-Pagan test manually.
}
\examples{
fit <- lm(mpg ~ wt + hp + disp, data = mtcars)
hetero_check(fit)

}
