% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{model_diagnostics}
\alias{model_diagnostics}
\title{Comprehensive Model Diagnostics}
\usage{
model_diagnostics(model)
}
\arguments{
\item{model}{A fitted linear model object.}
}
\value{
A data frame summarizing diagnostic results.
}
\description{
Summarizes key regression assumptions including
normality, heteroscedasticity, multicollinearity,
and influential observations.
}
\examples{
fit <- lm(mpg ~ wt + hp + disp, data = mtcars)
model_diagnostics(fit)

}
