% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multicol_check.R
\name{multicol_check}
\alias{multicol_check}
\title{Multicollinearity Check Using VIF}
\usage{
multicol_check(model)
}
\arguments{
\item{model}{A fitted linear model object.}
}
\value{
A data frame containing VIF values
and interpretation.
}
\description{
Computes Variance Inflation Factor (VIF)
for each predictor in a linear regression model.
}
\examples{
fit <- lm(mpg ~ wt + hp + disp, data = mtcars)
multicol_check(fit)

}
