% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_check.R
\name{normality_check}
\alias{normality_check}
\title{Normality Check for Model Residuals}
\usage{
normality_check(model)
}
\arguments{
\item{model}{A fitted linear model object.}
}
\value{
A data frame containing the test statistic,
p-value, and conclusion.
}
\description{
Performs a Shapiro-Wilk normality test on residuals from
a fitted linear regression model.
}
\examples{
fit <- lm(mpg ~ wt, data = mtcars)
normality_check(fit)

}
