% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_check.R
\name{outlier_check}
\alias{outlier_check}
\title{Outlier and Influence Check Using Cook's Distance}
\usage{
outlier_check(model)
}
\arguments{
\item{model}{A fitted linear model object.}
}
\value{
A data frame containing observation number,
Cook's Distance, and influence flag.
}
\description{
Identifies influential observations in a linear
regression model using Cook's Distance.
}
\examples{
fit <- lm(mpg ~ wt + hp, data = mtcars)
outlier_check(fit)

}
